Executing a Profile/Home While Already Executing
This commit should prevent a user from doubling up on motion profile executions. This is pertinent for home/touchoff/script execution. This is in regards to issue #39. This commit also address #26.

+++ src/ECM_gui/pump_dialog/window_pump_control.cpp
@@ -214,5 +214,5 @@ void Window_PumpControl::slot_PumpOperationalTimeout()
     common::EnvironmentTime currentTime;
     common::EnvironmentTime::CurrentTime(common::Devices::SYSTEMCLOCK,currentTime);
     double elapsedSeconds = (currentTime - startTime) / (1000.0 * 1000.0);
-    this->ui->lineEdit_OnTime->setText(QString::number(elapsedSeconds));
+    this->ui->lineEdit_OnTime->setText(QString::number((int)elapsedSeconds));
 }
